Summary Table Preparation

Returns a dataframe for the selected safety topic with the study ID, the number of patients in the selected arm, the number of patients in the arm with at least one occurrence of the adverse event, and the total exposure in patient years if the exposure-adjusted analysis is chosen.

Usage

data_table_prep(
  input_data,
  select_analysis,
  saf_topic,
  select_btrt,
  bool_pooled = FALSE,
  current_trial = FALSE,
  ae_summary = FALSE
)

Arguments

input_data

the raw summary level data

select_analysis

Incidence proportion or Exposure-adjusted AE rate

saf_topic

Selected safety topic to analyze/the adverse event of interest

select_btrt

selected background treatment

bool_pooled

Whether study data is pooled by study number

current_trial

information whether it is a historical (0) or current (1) trial

ae_summary

historical filtering for comparison, default FALSE

Value

a dataframe for the selected safety topic with the study ID, the number of patients in the selected arm, the number of patients in the arm with at least one occurrence of the adverse event, and the total exposure in patient years if the exposure-adjusted analysis is chosen