Automatic Fitting of Mixtures of Conjugate Distributions to a Sample

Fitting a series of mixtures of conjugate distributions to a sample using Expectation-Maximization. First, a Nc 1 component mixture is fitted, then a Nc 2 component mixture, and so on. The mixture providing the best AIC value is selected.

Usage

parametric_approx(select_analysis, map_prior)

Arguments

select_analysis

Incidence proportion or Exposure-adjusted AE rate

map_prior

an S3 object (list) of type gMAP for the incidence proportion analysis or a vector of the posterior MCMC samples for the exposure-adjusted analysis

Value

the best fitting mixture model, i.e., the model with the lowest AIC