Conjugate Posterior Analysis

Calculates the posterior distribution for given prior robust_map_prior, where the prior is a mixture of conjugate distributions. The posterior is then also a mixture of conjugate distributions. posterior MCMC samples from EXNEX analysis for exposure-adjusted analysis

Usage

posterior_dist(
  select_analysis,
  robust_map_prior,
  new_v1,
  new_v2,
  input_data,
  p_exch = NULL,
  nex_mean = NULL,
  nex_sd = NULL,
  adj_tau = NULL,
  seed = NULL
)

Arguments

select_analysis

Incidence proportion and Exposure-adjusted AE rate

robust_map_prior

mixture distribution output from robust_map()

new_v1

first new variable #n for proportion and #hazard rate for rates

new_v2

second new variable #r for proportion and unit information sd for rates

input_data

dataframe from data_table_prep() including the new trial

p_exch

probability of exchangeable component (use 1-robust_weight)

nex_mean

Mean for Nex part (use robust_mean)

nex_sd

Standard Deviation for Nex part

adj_tau

numeric value of tau from tau_adjust()

seed

for reproducibility

Value

posterior mixture distribution for incidence proportion analysis and