Summary statistics for robust prior, likelihood, and posterior

Summary statistics for robust prior, likelihood, and posterior in the New trial analysis tab

Usage

summary_stat_all_display(
  select_analysis,
  robust_map_object,
  ess_method,
  current_trial_data,
  post_dist,
  numerical = FALSE,
  seed
)

Arguments

select_analysis

Incidence proportion or Exposure-adjusted AE rate

robust_map_object

map prior; mixture distribution with a non-informative component from robust_map()

ess_method

ESS Method, currently only ELIR available

current_trial_data

information whether it is a historical (0) or current (1) trial

post_dist

posterior mixture distribution or MCMC samples from posterior_dist()

numerical

TRUE or FALSE, for return values in Dataframe or to be displayed

seed

a seed